Making everything private in the DefaultDocMapper
