+++ quickwit-doc-mapper/src/default_doc_mapper/default_mapper.rs
@@ -19,19 +19,23 @@
 
 use std::collections::BTreeSet;
 
+use anyhow::{bail, Context};
 use quickwit_proto::SearchRequest;
 use serde::{Deserialize, Serialize};
 use serde_json::{self, Value as JsonValue};
 use tantivy::query::Query;
-use tantivy::schema::{Field, Schema};
+use tantivy::schema::{Cardinality, Field, FieldType, Schema, STORED};
 use tantivy::Document;
+use tracing::info;
 
 use super::DefaultDocMapperBuilder;
-use crate::default_doc_mapper::mapping_tree::MappingNode;
+use crate::default_doc_mapper::mapping_tree::{build_mapping_tree, MappingNode, MappingTree};
 pub use crate::default_doc_mapper::QuickwitJsonOptions;
 use crate::query_builder::build_query;
 use crate::sort_by::{SortBy, SortOrder};
-use crate::{DocMapper, DocParsingError, ModeType, QueryParserError, DYNAMIC_FIELD_NAME};
+use crate::{
+    DocMapper, DocParsingError, ModeType, QueryParserError, DYNAMIC_FIELD_NAME, SOURCE_FIELD_NAME,
+};
 
 /// Specifies the name of the sort field and the sort order for an index.
 #[derive(Default, Serialize, Deserialize, Clone)]
@@ -84,28 +88,28 @@ impl Default for Mode {
 #[serde(try_from = "DefaultDocMapperBuilder", into = "DefaultDocMapperBuilder")]
 pub struct DefaultDocMapper {
     /// Field in which the source should be stored.
-    pub(crate) source_field: Option<Field>,
+    source_field: Option<Field>,
     /// Field in which the dynamically mapped fields should be stored.
-    pub(crate) dynamic_field: Option<Field>,
+    dynamic_field: Option<Field>,
     /// Default list of field names used for search.
-    pub(crate) default_search_field_names: Vec<String>,
+    default_search_field_names: Vec<String>,
     /// Timestamp field name.
-    pub(crate) timestamp_field_name: Option<String>,
+    timestamp_field_name: Option<String>,
     /// Sort field name and order.
-    pub(crate) sort_by: SortBy,
+    sort_by: SortBy,
     /// Root field mapping. It has an empty name so that we don't prefix all field names with it.
     /// See [FieldMappingEntry::root]
-    pub(crate) field_mappings: MappingNode,
+    field_mappings: MappingNode,
     /// Schema generated by the store source and field mappings parameters.
-    pub(crate) schema: Schema,
+    schema: Schema,
     /// List of field names used for tagging.
-    pub tag_field_names: BTreeSet<String>,
+    tag_field_names: BTreeSet<String>,
     /// Demux field name.
-    pub(crate) demux_field_name: Option<String>,
+    demux_field_name: Option<String>,
     /// List of required fields. Right now this is the list of fast fields.
-    pub(crate) required_fields: Vec<Field>,
+    required_fields: Vec<Field>,
     /// Defines how unmapped fields should be handle.
-    pub(crate) mode: Mode,
+    mode: Mode,
 }
 
 impl DefaultDocMapper {
@@ -119,6 +123,266 @@ impl DefaultDocMapper {
             }
         }
         Ok(())
+    }
+}
+
+/// Name of the raw tokenizer.
+const RAW_TOKENIZER_NAME: &str = "raw";
+
+fn validate_tag_fields(tag_fields: &[String], schema: &Schema) -> anyhow::Result<()> {
+    for tag_field in tag_fields {
+        let field = schema
+            .get_field(tag_field)
+            .ok_or_else(|| anyhow::anyhow!("Tag field `{}` does not exist.", tag_field))?;
+        let field_type = schema.get_field_entry(field).field_type();
+        match field_type {
+            FieldType::Str(options) => {
+                let tokenizer_opt = options
+                    .get_indexing_options()
+                    .map(|text_options| text_options.tokenizer());
+
+                if tokenizer_opt != Some(RAW_TOKENIZER_NAME) {
+                    bail!(
+                        "Tags collection is only allowed on text fields with the `raw` tokenizer."
+                    );
+                }
+            }
+            FieldType::Bytes(_) => {
+                bail!("Tags collection is not allowed on `bytes` fields.")
+            }
+            _ => (),
+        }
+    }
+    Ok(())
+}
+
+fn list_required_fields_for_node(node: &MappingNode) -> Vec<Field> {
+    node.children().flat_map(list_required_fields).collect()
+}
+
+fn list_required_fields(field_mappings: &MappingTree) -> Vec<Field> {
+    match field_mappings {
+        MappingTree::Leaf(leaf) => {
+            if leaf.get_type().is_fast_field() {
+                vec![leaf.field()]
+            } else {
+                Vec::new()
+            }
+        }
+        MappingTree::Node(node) => list_required_fields_for_node(node),
+    }
+}
+
+fn resolve_timestamp_field(
+    timestamp_field_name_opt: Option<&String>,
+    schema: &Schema,
+) -> anyhow::Result<()> {
+    if let Some(ref timestamp_field_name) = timestamp_field_name_opt {
+        let timestamp_field = schema
+            .get_field(timestamp_field_name)
+            .with_context(|| format!("Unknown timestamp field: `{}`", timestamp_field_name))?;
+
+        let timestamp_field_entry = schema.get_field_entry(timestamp_field);
+        if !timestamp_field_entry.is_fast() {
+            bail!(
+                "Timestamp field must be a fast field, please add the fast property to your field \
+                 `{}`.",
+                timestamp_field_name
+            )
+        }
+        match timestamp_field_entry.field_type() {
+            FieldType::I64(options) => {
+                if options.get_fastfield_cardinality() == Some(Cardinality::MultiValues) {
+                    bail!(
+                        "Timestamp field cannot be an array, please change your field `{}` from \
+                         an array to a single value.",
+                        timestamp_field_name
+                    )
+                }
+            }
+            _ => {
+                bail!(
+                    "Timestamp field must be of type i64, please change your field type `{}` to \
+                     i64.",
+                    timestamp_field_name
+                )
+            }
+        }
+    }
+    Ok(())
+}
+
+fn resolve_sort_field(
+    sort_by_config_opt: Option<SortByConfig>,
+    schema: &Schema,
+) -> anyhow::Result<SortBy> {
+    if let Some(sort_by_config) = sort_by_config_opt {
+        let sort_by_field = schema
+            .get_field(&sort_by_config.field_name)
+            .with_context(|| format!("Unknown sort by field: `{}`", sort_by_config.field_name))?;
+
+        let sort_by_field_entry = schema.get_field_entry(sort_by_field);
+        if !sort_by_field_entry.is_fast() {
+            bail!(
+                "Sort by field must be a fast field, please add the fast property to your field \
+                 `{}`.",
+                sort_by_config.field_name
+            )
+        }
+        return Ok(sort_by_config.into());
+    }
+    Ok(SortBy::DocId)
+}
+
+fn resolve_demux_field(
+    demux_field_name_opt: Option<&String>,
+    schema: &Schema,
+) -> anyhow::Result<()> {
+    if let Some(demux_field_name) = demux_field_name_opt {
+        let demux_field = schema
+            .get_field(demux_field_name)
+            .with_context(|| format!("Unknown demux field: `{}`", demux_field_name))?;
+
+        let demux_field_entry = schema.get_field_entry(demux_field);
+        if !demux_field_entry.is_fast() {
+            bail!(
+                "Demux field must be a fast field, please add the fast property to your field \
+                 `{}`.",
+                demux_field_name
+            )
+        }
+        if !demux_field_entry.is_indexed() {
+            bail!(
+                "Demux field must be indexed, please add the indexed property to your field `{}`.",
+                demux_field_name
+            )
+        }
+        match demux_field_entry.field_type() {
+            FieldType::U64(options) | FieldType::I64(options) => {
+                if options.get_fastfield_cardinality() == Some(Cardinality::MultiValues) {
+                    bail!(
+                        "Demux field cannot be an array, please change your field `{}` from an \
+                         array to a single value.",
+                        demux_field_name
+                    )
+                }
+            }
+            _ => {
+                bail!(
+                    "Demux field must be of type u64 or i64, please change your field type `{}` \
+                     to u64 or i64.",
+                    demux_field_name
+                )
+            }
+        }
+    }
+    Ok(())
+}
+
+impl TryFrom<DefaultDocMapperBuilder> for DefaultDocMapper {
+    type Error = anyhow::Error;
+
+    fn try_from(builder: DefaultDocMapperBuilder) -> Result<DefaultDocMapper, Self::Error> {
+        let mode = builder.mode()?;
+        let mut schema_builder = Schema::builder();
+        let field_mappings = build_mapping_tree(&builder.field_mappings, &mut schema_builder)?;
+        let source_field = if builder.store_source {
+            Some(schema_builder.add_text_field(SOURCE_FIELD_NAME, STORED))
+        } else {
+            None
+        };
+
+        let dynamic_field = if let Mode::Dynamic(json_options) = &mode {
+            Some(schema_builder.add_json_field(DYNAMIC_FIELD_NAME, json_options.clone()))
+        } else {
+            None
+        };
+
+        let schema = schema_builder.build();
+
+        // validate fast fields
+        validate_tag_fields(&builder.tag_fields, &schema)?;
+
+        // Resolve default search fields
+        let mut default_search_field_names = Vec::new();
+        for field_name in &builder.default_search_fields {
+            if default_search_field_names.contains(field_name) {
+                bail!("Duplicated default search field: `{}`", field_name)
+            }
+            schema
+                .get_field(field_name)
+                .with_context(|| format!("Unknown default search field: `{}`", field_name))?;
+            default_search_field_names.push(field_name.clone());
+        }
+
+        resolve_timestamp_field(builder.timestamp_field.as_ref(), &schema)?;
+        resolve_demux_field(builder.demux_field.as_ref(), &schema)?;
+        let sort_by = resolve_sort_field(builder.sort_by, &schema)?;
+
+        // Resolve tag fields
+        let mut tag_field_names: BTreeSet<String> = Default::default();
+        for tag_field_name in &builder.tag_fields {
+            if tag_field_names.contains(tag_field_name) {
+                bail!("Duplicated tag field: `{}`", tag_field_name)
+            }
+            schema
+                .get_field(tag_field_name)
+                .with_context(|| format!("Unknown tag field: `{}`", tag_field_name))?;
+            tag_field_names.insert(tag_field_name.clone());
+        }
+        if let Some(ref demux_field_name) = builder.demux_field {
+            if !tag_field_names.contains(demux_field_name) {
+                info!(
+                    "Demux field name `{}` is not in index config tags, add it automatically.",
+                    demux_field_name
+                );
+                tag_field_names.insert(demux_field_name.clone());
+            }
+        }
+
+        let required_fields = list_required_fields_for_node(&field_mappings);
+        Ok(DefaultDocMapper {
+            schema,
+            source_field,
+            dynamic_field,
+            default_search_field_names,
+            timestamp_field_name: builder.timestamp_field,
+            sort_by,
+            field_mappings,
+            tag_field_names,
+            required_fields,
+            demux_field_name: builder.demux_field,
+            mode,
+        })
+    }
+}
+
+impl From<DefaultDocMapper> for DefaultDocMapperBuilder {
+    fn from(default_doc_mapper: DefaultDocMapper) -> Self {
+        let sort_by_config = match &default_doc_mapper.sort_by {
+            SortBy::DocId => None,
+            SortBy::FastField { field_name, order } => Some(SortByConfig {
+                field_name: field_name.clone(),
+                order: *order,
+            }),
+        };
+        let demux_field = default_doc_mapper.demux_field_name();
+        let mode = default_doc_mapper.mode.mode_type();
+        let dynamic_mapping = match &default_doc_mapper.mode {
+            Mode::Dynamic(mapping_options) => Some(mapping_options.clone()),
+            _ => None,
+        };
+        Self {
+            store_source: default_doc_mapper.source_field.is_some(),
+            timestamp_field: default_doc_mapper.timestamp_field_name(),
+            field_mappings: default_doc_mapper.field_mappings.into(),
+            demux_field,
+            sort_by: sort_by_config,
+            tag_fields: default_doc_mapper.tag_field_names.into_iter().collect(),
+            default_search_fields: default_doc_mapper.default_search_field_names,
+            mode,
+            dynamic_mapping,
+        }
     }
 }
 
@@ -428,7 +692,7 @@ mod tests {
         let expected_msg = "Timestamp field must be a fast field, please add the fast property to \
                             your field `timestamp`."
             .to_string();
-        assert_eq!(builder.build().unwrap_err().to_string(), expected_msg);
+        assert_eq!(builder.try_build().unwrap_err().to_string(), expected_msg);
         Ok(())
     }
 
@@ -451,7 +715,7 @@ mod tests {
         let expected_msg = "Timestamp field cannot be an array, please change your field \
                             `timestamp` from an array to a single value."
             .to_string();
-        assert_eq!(builder.build().unwrap_err().to_string(), expected_msg);
+        assert_eq!(builder.try_build().unwrap_err().to_string(), expected_msg);
         Ok(())
     }
 
@@ -491,7 +755,7 @@ mod tests {
             ]
         }"#;
         let builder = serde_json::from_str::<DefaultDocMapperBuilder>(doc_mapper)?;
-        let doc_mapper = builder.build()?;
+        let doc_mapper = builder.try_build()?;
         let result = doc_mapper.doc_from_json(
             r#"{
             "image": "invalid base64 data"
@@ -527,7 +791,7 @@ mod tests {
         }"#;
 
         let builder = serde_json::from_str::<DefaultDocMapperBuilder>(doc_mapper).unwrap();
-        let doc_mapper = builder.build().unwrap();
+        let doc_mapper = builder.try_build().unwrap();
         let schema = doc_mapper.schema();
         const JSON_DOC_VALUE: &str = r#"{
             "city": "tokio",
@@ -579,7 +843,7 @@ mod tests {
         }"#;
         assert_eq!(
             serde_json::from_str::<DefaultDocMapperBuilder>(doc_mapper_one)?
-                .build()
+                .try_build()
                 .unwrap_err()
                 .to_string(),
             "Tags collection is only allowed on text fields with the `raw` tokenizer.".to_string(),
@@ -597,7 +861,7 @@ mod tests {
         }"#;
         assert_eq!(
             serde_json::from_str::<DefaultDocMapperBuilder>(doc_mapper_two)?
-                .build()
+                .try_build()
                 .unwrap_err()
                 .to_string(),
             "Tags collection is not allowed on `bytes` fields.".to_string(),
@@ -625,7 +889,7 @@ mod tests {
         let expected_msg = "Sort by field must be a fast field, please add the fast property to \
                             your field `timestamp`."
             .to_string();
-        assert_eq!(builder.build().unwrap_err().to_string(), expected_msg);
+        assert_eq!(builder.try_build().unwrap_err().to_string(), expected_msg);
         Ok(())
     }
 
@@ -778,7 +1042,7 @@ mod tests {
         let expected_msg = "Demux field must be indexed, please add the indexed property to your \
                             field `demux`."
             .to_string();
-        assert_eq!(builder.build().unwrap_err().to_string(), expected_msg);
+        assert_eq!(builder.try_build().unwrap_err().to_string(), expected_msg);
         Ok(())
     }
 
@@ -796,7 +1060,7 @@ mod tests {
             ]
         }"#;
         let builder = serde_json::from_str::<DefaultDocMapperBuilder>(doc_mapper).unwrap();
-        let default_doc_mapper = builder.build().unwrap();
+        let default_doc_mapper = builder.try_build().unwrap();
         assert!(default_doc_mapper.source_field.is_none());
         let schema = default_doc_mapper.schema();
         let field = schema.get_field("my-field").unwrap();
